Add files via upload
The project uses computer vision and mediapipe library to control systems volume based on hand gesture. The code continuously takes video frames as the input and detects hand landmarks. It calculates the distance between the fingers, maps the distance to control the system volume and provides real time visual representation. The code uses 'pycaw' library to manage the audio which is compatible only with windows devices.